Refactor with only simapro csv import
